Microsoft edge update workaround

+++ run_zwift.sh
@@ -51,10 +51,19 @@ echo "Killing unnecessary applications"
 # container to exit.  See https://github.com/netbrain/zwift/issues/210
 pkill ZwiftLauncher || true
 pkill ZwiftWindowsCra
-pkill -f MicrosoftEdgeUpdate
 
 if [ -z "$ZWIFT_NO_GAMEMODE" ]; then
     /usr/games/gamemoderun wineserver -w
 else
     wineserver -w
+fi
+
+echo "Waiting for zwift to stop..."
+while pgrep -f ZwiftApp.exe &> /dev/null; do
+    sleep 5
+done
+echo "Zwift exited"
+if pgrep -f MicrosoftEdgeUpdate &> /dev/null; then
+    echo "Killing Microsoft Edge Update"
+    pkill -f MicrosoftEdgeUpdate
 fi
